Move testInvalidCredentialFromAccessTokenProvider to GoogleHadoopFileSystemTestBase.
This test is not GHFS-implementation specific.

	Change on 2018/05/23 by idv <idv@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197825150
